+++ Source/OpenTK/Compute/CL12/CL12.Enums.cs
@@ -1734,7 +1734,7 @@ namespace OpenTK.Compute.CL12
     /// <summary>
     /// Not used directly.
     /// </summary>
-    public enum CommandQueueFlags : int
+    public enum CommandQueueFlags : long
     {
         /// <summary>
         /// Original was CL_QUEUE_OUT_OF_ORDER_EXEC_MODE_ENABLE = (1 << 0)
@@ -2370,7 +2370,7 @@ namespace OpenTK.Compute.CL12
     /// <summary>
     /// Not used directly.
     /// </summary>
-    public enum DeviceTypeFlags : int
+    public enum DeviceTypeFlags : long
     {
         /// <summary>
         /// Original was CL_DEVICE_TYPE_DEFAULT = (1 << 0)
