+++ Source/OpenTK/Compute/CL20/CL20.Enums.cs
@@ -1946,7 +1946,7 @@ namespace OpenTK.Compute.CL20
     /// <summary>
     /// Not used directly.
     /// </summary>
-    public enum CommandQueueFlags : int
+    public enum CommandQueueFlags : long
     {
         /// <summary>
         /// Original was CL_QUEUE_OUT_OF_ORDER_EXEC_MODE_ENABLE = (1 << 0)
@@ -2693,7 +2693,7 @@ namespace OpenTK.Compute.CL20
     /// <summary>
     /// Not used directly.
     /// </summary>
-    public enum DeviceTypeFlags : int
+    public enum DeviceTypeFlags : long
     {
         /// <summary>
         /// Original was CL_DEVICE_TYPE_DEFAULT = (1 << 0)
